Refactor database models: remove old Article, User, and Session schemas; add new implementations in article.js, user.js, and session.js; update import paths in service and middleware files.
