+++ src/middlewares/authenticate.js
@@ -1,6 +1,6 @@
 import createHttpError from 'http-errors';
-import { Session } from '../db/models/123.js';
-import User from '../db/models/666.js';
+import Session from '../db/models/session.js';
+import User from '../db/models/user.js';
 
 export const authenticate = async (req, res, next) => {
   const authHeader = req.get('Authorization');
