+++ src/services/auth/logoutUserService.js
@@ -1,4 +1,4 @@
-import { Session } from '../../db/models/123.js';
+import { Session } from '../../db/models/session.js';
 
 export const logoutUser = async (sessionId, sessionToken) => {
   const deletedSession = await Session.findOneAndDelete({
