+++ src/services/auth/refreshSession.js
@@ -1,5 +1,5 @@
 import createHttpError from 'http-errors';
-import { Session } from '../../db/models/123.js';
+import { Session } from '../../db/models/session.js';
 import { generateSessionTokens } from '../../utils/generateSessionTokens.js';
 
 export const refreshSession = async (sessionId, sessionToken) => {
