+++ src/services/users/allUsersInfoService.js
@@ -1,4 +1,4 @@
-import User from '../../db/models/User.js';
+import User from '../../db/models/user.js';
 
 export const allUsersInfoService = async () => {
   return await User.find().select('-password').select('-email');
