+++ src/services/users/getUserArticlesServise.js
@@ -1,4 +1,4 @@
-import Article from '../../db/models/Article.js';
+import Article from '../../db/models/article.js';
 
 export const getUserArticlesServise = async (id) => {
   return await Article.find({ ownerId: id });
