+++ src/services/users/saveArticleToUserServise.js
@@ -1,4 +1,4 @@
-import User from '../../db/models/User.js';
+import User from '../../db/models/user.js';
 import createHttpError from 'http-errors';
 
 export const saveArticleToUserServise = async (userId, articleId) => {
